moving ecal cuda stuff to be plugins only
