+++ EventFilter/EcalRawToDigi/plugins/BuildFile.xml
@@ -18,6 +18,6 @@
 <use name="TrackingTools/Records"/>
 <use name="TrackingTools/GeomPropagators"/>
 <use name="TrackingTools/TrajectoryState"/>
-<library file="*.cc" name="EventFilterEcalRawToDigiPlugins">
+<library file="*.cc *.cu" name="EventFilterEcalRawToDigiPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
