+++ EventFilter/EcalRawToDigi/plugins/DeclsForKernels.h
@@ -1,5 +1,5 @@
-#ifndef EventFilter_EcalRawToDigi_interface_DeclsForKernels_h
-#define EventFilter_EcalRawToDigi_interface_DeclsForKernels_h
+#ifndef EventFilter_EcalRawToDigi_plugins_DeclsForKernels_h
+#define EventFilter_EcalRawToDigi_plugins_DeclsForKernels_h
 
 #include <vector>
 
@@ -107,4 +107,4 @@ namespace ecal {
   }  // namespace raw
 }  // namespace ecal
 
-#endif  // EventFilter_EcalRawToDigi_interface_DeclsForKernels_h
+#endif  // EventFilter_EcalRawToDigi_plugins_DeclsForKernels_h
