+++ EventFilter/EcalRawToDigi/plugins/UnpackGPU.h
@@ -1,7 +1,7 @@
-#ifndef EventFilter_EcalRawToDigi_interface_UnpackGPU_h
-#define EventFilter_EcalRawToDigi_interface_UnpackGPU_h
+#ifndef EventFilter_EcalRawToDigi_plugins_UnpackGPU_h
+#define EventFilter_EcalRawToDigi_plugins_UnpackGPU_h
 
-#include "EventFilter/EcalRawToDigi/interface/DeclsForKernels.h"
+#include "EventFilter/EcalRawToDigi/plugins/DeclsForKernels.h"
 
 namespace ecal {
   namespace raw {
@@ -20,4 +20,4 @@ namespace ecal {
   }  // namespace raw
 }  // namespace ecal
 
-#endif  // EventFilter_EcalRawToDigi_interface_UnpackGPU_h
+#endif  // EventFilter_EcalRawToDigi_plugins_UnpackGPU_h
